requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.8.1
astroid==3.1.0
attrs==23.2.0
autopep8==2.0.4
bleach==6.1.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect==1.2.0
decorator==5.1.1
dicttoxml==1.7.16
dill==0.3.8
Django==5.0.3
django-redis==5.4.0
djangorestframework==3.15.1
docutils==0.20.1
frozenlist==1.4.1
future==1.0.0
geocoder==1.38.1
geoip2==4.8.0
h11==0.14.0
idna==3.6
importlib_metadata==7.0.2
ip2geotools==0.1.6
IP2Location==8.10.2
isort==5.13.2
jaraco.classes==3.3.1
keyring==24.3.1
lazy-object-proxy==1.10.0
lxml==5.1.0
markdown-it-py==3.0.0
maxminddb==2.5.2
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.2.0
multidict==6.0.5
nh3==0.2.15
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
pillow==10.2.0
pip-review==1.3.0
pkginfo==1.10.0
platformdirs==4.2.0
psycopg2==2.9.9
pycodestyle==2.11.1
pycparser==2.21
Pygments==2.17.2
pylint==3.1.0
pyparsing==3.1.2
pypng==0.20220715.0
pyquery==2.0.0
PySocks==1.7.1
pywin32-ctypes==0.2.2
qrcode==7.4.2
ratelim==0.1.6
readme_renderer==43.0
redis==5.0.4
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
selenium==4.18.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
tomlkit==0.12.4
tqdm==4.66.2
trio==0.24.0
trio-websocket==0.11.1
twine==5.0.0
typed-ast==1.5.5
typing==3.7.4.3
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.18.0
